Move settings page to component
